TimeTable Complete
The app is now fully functioning with ability to create, edit and
display semesters and courses with a wizard and a double week view. The
basis of the app is complete, all that is left is to make it pretty, and
add the other views etc. Features ++ from here on in.